refactored enter_exit
